test(journeys): convert space suite
