+++ test/journeys/specs/space/data-api.js
@@ -1,27 +1,22 @@
-import {assert} from 'chai';
-
 import testUsers from '@ciscospark/test-helper-test-users';
 import '@ciscospark/plugin-logger';
 import CiscoSpark from '@ciscospark/spark-core';
 import '@ciscospark/internal-plugin-conversation';
 
-import {updateJobStatus} from '../../../lib/test-helpers';
-import waitForPromise from '../../../lib/wait-for-promise';
+import {jobNames, updateJobStatus} from '../../lib/test-helpers';
+import waitForPromise from '../../lib/wait-for-promise';
+import {switchToMeet} from '../../lib/test-helpers/space-widget/main';
+import {elements, declineIncomingCallTest, hangupBeforeAnswerTest, hangupDuringCallTest} from '../../lib/test-helpers/space-widget/meet';
 import {
-  canDeleteMessage,
-  deleteMessage,
-  flagMessage,
-  messageTests,
-  removeFlagMessage,
   sendMessage,
   verifyMessageReceipt
-} from '../../../lib/test-helpers/space-widget/messaging';
+} from '../../lib/test-helpers/space-widget/messaging';
 
-describe('Widget Space: Data API', () => {
+describe('Space Widget Data API Tests', () => {
   const browserLocal = browser.select('browserLocal');
   const browserRemote = browser.select('browserRemote');
   const spaceWidget = '.ciscospark-space-widget';
-  const jobName = 'react-widget-space-dataApi';
+  const jobName = jobNames.space;
 
   let allPassed = true;
   let docbrown, lorraine, marty;
@@ -140,99 +135,27 @@ describe('Widget Space: Data API', () => {
       sendMessage(local, remote, martyText2);
       verifyMessageReceipt(remote, local, martyText2);
     });
+  });
 
-    describe('message actions', () => {
-      describe('message flags', () => {
-        const message = 'Do you really think this is a good idea?';
-        before('create a message to flag', () => {
-          sendMessage(remote, local, message);
-          verifyMessageReceipt(local, remote, message);
-        });
-
-        it('should be able to flag a message', () => {
-          flagMessage(local, message);
-        });
-
-        it('should be able to unflag a message', () => {
-          removeFlagMessage(local, message);
-        });
-      });
-
-      describe('delete message', () => {
-        it('should be able to delete a message from self', () => {
-          const message = 'There is no spoon!';
-          sendMessage(local, remote, message);
-          verifyMessageReceipt(remote, local, message);
-          deleteMessage(local, message);
-        });
-
-        it('should not be able to delete a message from others', () => {
-          const message = 'Hey you guys!';
-          sendMessage(remote, local, message);
-          verifyMessageReceipt(local, remote, message);
-          assert.isFalse(canDeleteMessage(local, message));
-        });
+  describe('meet widget', () => {
+    describe('pre call experience', () => {
+      it('has a call button', () => {
+        switchToMeet(browserLocal);
+        browserLocal.waitForVisible(elements.callButton);
       });
     });
 
-    // Skipped due to file transfer instability (2018-07-09)
-    describe.skip('File Transfer Tests', () => {
-      it('sends message with png attachment', () => {
-        messageTests.sendFileTest(local, remote, 'png-sample.png');
+    describe('during call experience', () => {
+      it('can hangup before answer', () => {
+        hangupBeforeAnswerTest(browserLocal, browserRemote);
       });
 
-      it('verifies png-sample is in files tab', () => {
-        messageTests.filesTabTest(local, remote, 'png-sample.png');
-      });
-    });
-
-    describe('markdown messaging', () => {
-      it('sends message with bold text', () => {
-        messageTests.markdown.bold(remote, local);
+      it('can decline an incoming call', () => {
+        declineIncomingCallTest(browserLocal, browserRemote, true);
       });
 
-      it('sends message with italic text', () => {
-        messageTests.markdown.italic(local, remote);
-      });
-
-      it('sends message with a blockquote', () => {
-        messageTests.markdown.blockquote(remote, local);
-      });
-
-      it('sends message with numbered list', () => {
-        messageTests.markdown.orderedList(local, remote);
-      });
-
-      it('sends message with bulleted list', () => {
-        messageTests.markdown.unorderedList(remote, local);
-      });
-
-      it('sends message with heading 1', () => {
-        messageTests.markdown.heading1(local, remote);
-      });
-
-      it('sends message with heading 2', () => {
-        messageTests.markdown.heading2(remote, local);
-      });
-
-      it('sends message with heading 3', () => {
-        messageTests.markdown.heading3(local, remote);
-      });
-
-      it('sends message with horizontal line', () => {
-        messageTests.markdown.hr(remote, local);
-      });
-
-      it('sends message with link', () => {
-        messageTests.markdown.link(local, remote);
-      });
-
-      it('sends message with inline code', () => {
-        messageTests.markdown.inline(remote, local);
-      });
-
-      it('sends message with codeblock', () => {
-        messageTests.markdown.codeblock(local, remote);
+      it('can hangup in call', () => {
+        hangupDuringCallTest(browserLocal, browserRemote, true);
       });
     });
   });
